Moved routes to their own file. Started working on app routes. Added styling to landing page
