+++ public/assets/css/style.css
@@ -17,40 +17,24 @@ body {
 }
 
 #header-text{
-    font-family: 'Josefin Sans', sans-serif;
+    font-family: "Gotham-Light", "Gotham Light", "Gotham", Helvetica, Roboto, Arial, sans-serif;
     font-size: 80px;
-    /*align-content: center;*/
+    margin-top: 100px;
+    font-weight: 100;
+    letter-spacing: 4px;
 }
 
 #sub-text{
     font-family: 'Josefin Sans', sans-serif;
     font-size: 45px;
-    /*align-content: center;*/
+    margin-bottom: 100px;
 }
-#video-bg-elem {
-    /*height: 100%;*/
-    width: 100%;
-}
-
-body{
-    background-color: blue;
-}
-
-.background-wrap{
-    position: fixed;
-    z-index: -1000;
-    width: 100%;
-    height: 100%;
-    overflow: hidden;
-    top: 0;
-    left: 0;
 
   /* Set the fixed height of the footer here */
 .footer {
   position: absolute;
   bottom: 0;
   width: 100%;
-  padding-top: 10px;
   /* Set the fixed height of the footer here */
   height: 60px;
   background-color: #f5f5f5;
@@ -80,8 +64,9 @@ code {
   font-size: 80%;
 }
 
-
-
+/*******************************
+ *  STYLES FOR REGISTER MODAL  *
+ *******************************/
 .form-group.required .required-field:after {
   content: "*";
   color: red;
@@ -90,4 +75,58 @@ code {
 label.input-label {
   width: 96%;
   font-weight: 400;
+}
+
+/*********************
+ *  LANDING CONTENT  *
+ *********************/
+#landing-content {
+  margin-top: -20px;
+  padding-bottom: 80px;
+  background-color: black;
+  color: white;
+  background: linear-gradient(rgba(99, 100, 102, 0.7), rgba(48, 49, 51, 0.7)), url(../images/burial.gif) center no-repeat;
+  -webkit-background-size: cover;
+  -moz-background-size: cover;
+  -o-background-size: cover;
+  background-size: cover;
+
+}
+
+#landing-content form {
+  margin-top: 20px;
+}
+
+#landing-bottom-content .container {
+  margin-top: 40px;
+  margin-bottom: 110px;
+}
+
+#landing-bottom-content .col-sm-5 {
+  padding: 20px 0px;
+}
+
+#landing-bottom-content img {
+  width: 100%;
+  -moz-transform: scaleX(-1);
+  -o-transform: scaleX(-1);
+  -webkit-transform: scaleX(-1);
+  transform: scaleX(-1);
+  filter: FlipH;
+  -ms-filter: "FlipH";
+}
+
+#landing-bottom-content .column-text {
+  ont-family: "Gotham-Light", "Gotham Light", "Gotham", Helvetica, Roboto, Arial, sans-serif;
+}
+
+#landing-bottom-content .column-text div {
+  font-size: 16pt;
+}
+
+.message-display {
+  position: fixed;
+  width: 75%;
+  left: 50%;
+  margin-left: -37.5%;
 }
