+++ views/partials/header.ejs
@@ -125,14 +125,6 @@
                   </label>
                 </div>
               </div>
-              <!--<div class="form-group required">-->
-                <!--<label for="register-bday" class="col-sm-3 control-label">Birth</label>-->
-                <!--<div class="col-sm-8 required-field">-->
-                  <!--<label class="input-label">-->
-                    <!--<input type="date" required class="form-control" id="register-bday" name="bday">-->
-                  <!--</label>-->
-                <!--</div>-->
-              <!--</div>-->
               <div class="form-group">
                 <div class="col-sm-offset-3 col-sm-8">
                   <button type="submit" id="register-submit-button" class="btn btn-default pull-right">Register</button>
@@ -145,7 +137,7 @@
       </div>
     </div>
 
-    <div class="container">
+    <div class="container message-display">
       <% if(success && success.length > 0) { %>
       <div class="alert alert-success alert-dismissible" role="alert">
         <button type="button" class="close" data-dismiss="alert" aria-label="Close">
