+++ views/second.ejs
@@ -1,0 +1,13 @@
+<% include partials/header %>
+
+<div class="container">
+  <div class="row">
+    <div class="col-lg-12">
+      <div class="content">
+        SECOND WISHES PAGE
+      </div>
+    </div>
+  </div>
+</div>
+
+<% include partials/footer %>
